Create README.md
This is the story of a goat. A long time ago, there lived a very playful goat. The goat lived with its owners, Mr. and Mrs. Singh. They pampered the goat like their own child. But the goat was very dowdy and loved woollen coats. Mr. and Mrs. Singh were very worried about it. They scolded the goat every day for its behaviour.

One day Mr. and Mrs. Singh had to go to the market to buy fruits and vegetables. Before leaving, Mrs. Singh had put Mr. Singh's favourite woollen jacket on a hanger to dry. Both of them warned the goat not to chew the jacket. Once they left, it started raining heavily. The jacket fell to the ground due to the strong wind.

When the goat was picking up the jacket, unknowingly a piece of the jacket came into its mouth. Before it could understand anything, it chewed the jacket by mistake. When the rain stopped, the goat realised what it had done. Finally, she decided to keep quiet

When Mr. and Mrs. Singh returned, they found that the jacket was missing. They searched hard, but couldn't find the woollen jacket. Mrs. Singh noticed a strand of wool stuck around the chin of their beloved goat. Mrs. Singh angrily shouted at the goat and kicked it out of their house. The poor goat was kicked out of the house without even being fed. It vowed never to chew wool again